First version of the completed database updater
